Multiple Columns and Complex Conditions:
5)Given a DataFrame df with columns temperature (in Celsius) and humidity, create a new column
weather_type with the following conditions:
If temperature is greater than or equal to 30 and humidity is greater than or equal to 60, set
weather_type to "Hot and Humid".
If temperature is less than 30 and humidity is greater than or equal to 60, set weather_type to
"Warm and Humid".
If temperature is greater than or equal to 30 and humidity is less than 60, set weather_type to "Hot
and Dry".
If temperature is less than 30 and humidity is less than 60, set weather_type to "Moderate".