Combining Multiple Conditions:
7)Given a DataFrame df with columns sales, expenses, and profit, create a new column profit_status
based on the following conditions:
If profit is greater than 0, set profit_status to "Profitable".
If profit is equal to 0, set profit_status to "Break-even".
If profit is less than 0 and sales is greater than 0, set profit_status to "Loss-making".
If profit is less than 0 and sales is 0, set profit_status to "No Sales".